+++ __tests__/seeds/commentSeeds.js
@@ -2,14 +2,46 @@ const comment_input = [
   `/*!
  * @file FileName.js
  */`,
+  `/**
+ * @file FileName.js
+ */`,
+  `/*!
+ * \\file FileName.js
+ */`,
   `/*!
  * @file FileName.js
  * @brief file description
+ */`,
+  `/**
+ * @file FileName.js
+ * @brief file description
+ */`,
+  `/*!
+ * \\file FileName.js
+ * \\brief file description
  */`,
 ];
 
 const comment_output = [
   `/*!
+ * @file FileName.js
+ */
+`,
+  `/*!
+ * @file FileName.js
+ */
+`,
+  `/*!
+ * @file FileName.js
+ */
+`,
+  `/*!
+ * @brief file description
+ * @file FileName.js
+ */
+`,
+  `/*!
+ * @brief file description
  * @file FileName.js
  */
 `,