+++ __tests__/seeds/variableSeeds.js
@@ -1,7 +1,15 @@
 const variables_input = [
   `var x;`,
+  `let x;`,
+  `const x;`,
   `//! variable description
 var x;
+`,
+  `//! variable description
+let x;
+`,
+  `//! variable description
+const x;
 `,
   `//! type:String variable description
 var x;
@@ -10,6 +18,16 @@ var x;
 
 const variables_output = [
   `var x;
+`,
+  `var x;
+`,
+  `var x;
+`,
+  `//! variable description
+var x;
+`,
+  `//! variable description
+var x;
 `,
   `//! variable description
 var x;